# Conflicts:
#	app/Http/Controllers/V2/InstanceController.php
#	app/Http/Requests/V2/Instance/CreateRequest.php
#	app/Models/V2/Instance.php
#	app/Resources/V2/InstanceResource.php
#	docs/v2/components/schemas/instance.yaml

+++ app/Http/Controllers/V2/InstanceController.php
@@ -2,11 +2,15 @@
 
 namespace App\Http\Controllers\V2;
 
-use App\Http\Requests\V2\CreateInstanceRequest;
-use App\Http\Requests\V2\UpdateInstanceRequest;
-use App\Models\V2\Instance;
-use App\Resources\V2\InstanceResource;
 use Illuminate\Http\Request;
+use App\Http\Requests\V2\Instance\CreateRequest;
+use App\Http\Requests\V2\Instance\DeployRequest;
+use App\Http\Requests\V2\Instance\UpdateRequest;
+use App\Models\V2\Instance;
+use App\Resources\V2\CredentialResource;
+use App\Resources\V2\InstanceResource;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use UKFast\DB\Ditto\QueryTransformer;
 
 /**
@@ -23,7 +27,6 @@ class InstanceController extends BaseController
      */
     public function index(Request $request, QueryTransformer $queryTransformer)
     {
-
         $collection = Instance::forUser($request->user);
 
         $queryTransformer->config(Instance::class)
@@ -47,26 +50,34 @@ class InstanceController extends BaseController
     }
 
     /**
-     * @param \App\Http\Requests\V2\CreateInstanceRequest $request
+     * @param  CreateRequest  $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function store(CreateInstanceRequest $request)
+    public function store(CreateRequest $request)
     {
-        $instance = new Instance($request->only(['network_id', 'name', 'vpc_id', 'availability_zone_id']));
+        $instance = new Instance($request->only(['name', 'vpc_id', 'availability_zone_id', 'locked']));
+        if (!$request->has('locked')) {
+            $instance->locked = false;
+        }
         $instance->save();
         $instance->refresh();
         return $this->responseIdMeta($request, $instance->getKey(), 201);
     }
 
     /**
-     * @param UpdateInstanceRequest $request
+     * @param UpdateRequest $request
      * @param string $instanceId
      * @return \Illuminate\Http\JsonResponse
      */
-    public function update(UpdateInstanceRequest $request, string $instanceId)
+    public function update(UpdateRequest $request, string $instanceId)
     {
         $instance = Instance::forUser(app('request')->user)->findOrFail($instanceId);
-        $instance->fill($request->only(['vpc_id', 'name']));
+        if (!$this->isAdmin &&
+            (!$request->has('locked') || $request->get('locked') !== false) &&
+            $instance->locked === true) {
+            return $this->isLocked();
+        }
+        $instance->fill($request->only(['name', 'vpc_id', 'locked', 'availability_zone_id']));
         $instance->save();
         return $this->responseIdMeta($request, $instance->getKey(), 200);
     }
@@ -79,7 +90,71 @@ class InstanceController extends BaseController
     public function destroy(Request $request, string $instanceId)
     {
         $instance = Instance::forUser($request->user)->findOrFail($instanceId);
+        if (!$this->isAdmin && $instance->locked === true) {
+            return $this->isLocked();
+        }
         $instance->delete();
         return response()->json([], 204);
+    }
+
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    private function isLocked(): JsonResponse
+    {
+        return JsonResponse::create([
+            'errors' => [
+                'title'  => 'Forbidden',
+                'detail' => 'The specified instance is locked',
+                'status' => 403,
+            ]
+        ], 403);
+    }
+
+    /**
+     * @param  Request  $request
+     * @param  string  $instanceId
+     *
+     * @return AnonymousResourceCollection|\Illuminate\Support\HigherOrderTapProxy|mixed
+     */
+    public function credentials(Request $request, string $instanceId)
+    {
+        return CredentialResource::collection(
+            Instance::forUser($request->user)
+                ->findOrFail($instanceId)
+                ->credentials()
+                ->paginate($request->input('per_page', env('PAGINATION_LIMIT')))
+        );
+    }
+
+    public function deploy(DeployRequest $request, string $instanceId)
+    {
+        $instance = Instance::forUser(app('request')->user)->findOrFail($instanceId);
+        if (!$instance) {
+            return response()->json([], 404);
+        }
+
+        $data = [
+            'instance_id' => $instance->id,
+            'vpc_id' => $instance->vpc->id,
+            'volume_capacity' => $request->input('volume_capacity', config('volume.capacity.min')),
+            'network_id' => $request->input('network_id'),
+            'floating_ip_id' => $request->input('floating_ip_id'),
+            'appliance_data' => $request->input('appliance_data'),
+        ];
+
+        // Create the chained jobs for deployment
+        $this->dispatch((new \App\Jobs\Instance\Deploy\Deploy($data))->chain([
+            new \App\Jobs\Instance\Deploy\UpdateNetworkAdapter($data),
+            new \App\Jobs\Instance\Deploy\PowerOn($data),
+            new \App\Jobs\Instance\Deploy\WaitOsCustomisation($data),
+            new \App\Jobs\Instance\Deploy\PrepareOsUsers($data),
+            new \App\Jobs\Instance\Deploy\OsCustomisation($data),
+            new \App\Jobs\Instance\Deploy\PrepareOsDisk($data),
+            new \App\Jobs\Instance\Deploy\RunApplianceBootstrap($data),
+            new \App\Jobs\Instance\Deploy\RunBootstrapScript($data),
+        ]));
+
+        return response()->json([], 202);
     }
 }
